Split the bot into three different files to allow the easier implementation of functionalities
